services/main_orchestrator/requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg2-binary
redis
pydantic
aiohttp
requests
